feat: 플로팅 버튼 안뜨던 부분 수정
